fix unit test failure
