+++ src/vs/workbench/services/layout/browser/layoutService.ts
@@ -9,6 +9,11 @@ import { ILayoutService } from 'vs/platform/layout/browser/layoutService';
 import { Part } from 'vs/workbench/browser/part';
 import { IDimension } from 'vs/base/browser/dom';
 import { Direction } from 'vs/base/browser/ui/grid/grid';
+import { isMacintosh, isNative, isWeb } from 'vs/base/common/platform';
+import { isAuxiliaryWindow } from 'vs/base/browser/window';
+import { CustomTitleBarVisibility, TitleBarSetting, getMenuBarVisibility, hasCustomTitlebar, hasNativeTitlebar } from 'vs/platform/window/common/window';
+import { isFullscreen, isWCOEnabled } from 'vs/base/browser/browser';
+import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
 
 export const IWorkbenchLayoutService = refineServiceDecorator<ILayoutService, IWorkbenchLayoutService>(ILayoutService);
 
@@ -302,4 +307,85 @@ export interface IWorkbenchLayoutService extends ILayoutService {
 	 * Returns the next visible view part in a given direction in the main window.
 	 */
 	getVisibleNeighborPart(part: Parts, direction: Direction): Parts | undefined;
+}
+
+export function shouldShowCustomTitleBar(configurationService: IConfigurationService, window: Window, menuBarToggled?: boolean): boolean {
+
+	if (!hasCustomTitlebar(configurationService)) {
+		return false;
+	}
+
+	const inFullscreen = isFullscreen(window);
+	const nativeTitleBarEnabled = hasNativeTitlebar(configurationService);
+
+	const showCustomTitleBar = configurationService.getValue<CustomTitleBarVisibility>(TitleBarSetting.CUSTOM_TITLE_BAR_VISIBILITY);
+	if (showCustomTitleBar === CustomTitleBarVisibility.NEVER && nativeTitleBarEnabled || showCustomTitleBar === CustomTitleBarVisibility.WINDOWED && inFullscreen) {
+		return false;
+	}
+
+	if (!isTitleBarEmpty(configurationService)) {
+		return true;
+	}
+
+	// Hide custom title bar when native title bar enabled and custom title bar is empty
+	if (nativeTitleBarEnabled) {
+		return false;
+	}
+
+	// macOS desktop does not need a title bar when full screen
+	if (isMacintosh && isNative) {
+		return !inFullscreen;
+	}
+
+	// non-fullscreen native must show the title bar
+	if (isNative && !inFullscreen) {
+		return true;
+	}
+
+	// if WCO is visible, we have to show the title bar
+	if (isWCOEnabled() && !inFullscreen) {
+		return true;
+	}
+
+	// remaining behavior is based on menubar visibility
+	const menuBarVisibility = !isAuxiliaryWindow(window) ? getMenuBarVisibility(configurationService) : 'hidden';
+	switch (menuBarVisibility) {
+		case 'classic':
+			return !inFullscreen || !!menuBarToggled;
+		case 'compact':
+		case 'hidden':
+			return false;
+		case 'toggle':
+			return !!menuBarToggled;
+		case 'visible':
+			return true;
+		default:
+			return isWeb ? false : !inFullscreen || !!menuBarToggled;
+	}
+}
+
+function isTitleBarEmpty(configurationService: IConfigurationService): boolean {
+	// with the command center enabled, we should always show
+	if (configurationService.getValue<boolean>(LayoutSettings.COMMAND_CENTER)) {
+		return false;
+	}
+
+	// with the activity bar on top, we should always show
+	if (configurationService.getValue(LayoutSettings.ACTIVITY_BAR_LOCATION) === ActivityBarPosition.TOP) {
+		return false;
+	}
+
+	// with the editor actions on top, we should always show
+	const editorActionsLocation = configurationService.getValue<EditorActionsLocation>(LayoutSettings.EDITOR_ACTIONS_LOCATION);
+	const editorTabsMode = configurationService.getValue<EditorTabsMode>(LayoutSettings.EDITOR_TABS_MODE);
+	if (editorActionsLocation === EditorActionsLocation.TITLEBAR || editorActionsLocation === EditorActionsLocation.DEFAULT && editorTabsMode === EditorTabsMode.NONE) {
+		return false;
+	}
+
+	// with the layout actions on top, we should always show
+	if (configurationService.getValue<boolean>(LayoutSettings.LAYOUT_ACTIONS)) {
+		return false;
+	}
+
+	return true;
 }
